bunch of stuff
removed fire extinguisher, created and closed off door that is meant to explode, can't quite get it working. Set up some enemies to teleport in later on once I figure teleporters out, added geometry to every combat section except for the large ramp to the exit room. Geometry isn't final and is just there to work off of.  Added a secret in the generator room, but no way to escape yet, planning on using trains on a trigger.