finished first draft
The first draft of the gameplay blockout is finished, I added the groundwork for a new secret but I'm not confident in the full idea yet so I'm going to leave it undiscoverable and keep workshopping. I think some of the combat might be unrefined, and I'm not certain about item placement, but I plan on playtesting with a group tomorrow, so I will refine after that.